# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ipmoney API
  version: 0.1.0
  license:
    name: Proprietary
    url: https://ipmoney.com
  description: |
    Ipmoney 专利交易平台（小程序 + 用户 H5 + 后台）对外 API（P0）。

    业务约束（P0）：
    - 标的范围：仅 CN 专利（数据以用户上传/后台录入为准）
    - 资金路径：订金与尾款均线上支付并托管
    - 合同签署：线下签署完成后由客服/运营确认里程碑解锁尾款支付
    - 佣金：卖家承担；放款时从卖家应收中扣除
servers:
  - url: https://api.ipmoney.com/v1
...
security:
  - BearerAuth: []
paths:
...
  /public/achievements/{achievementId}:
    get:
      tags: [Achievements]
      summary: 获取成果详情（公开）
      operationId: getPublicAchievementById
      security: []
      parameters:
        - $ref: "#/components/parameters/AchievementId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AchievementPublic"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /search/achievements:
    get:
      tags: [Search]
      summary: 成果展示检索
      operationId: searchAchievements
      security: []
      parameters:
        - $ref: "#/components/parameters/Q"
        - $ref: "#/components/parameters/RegionCode"
        - $ref: "#/components/parameters/IndustryTags"
        - $ref: "#/components/parameters/CooperationModes"
        - $ref: "#/components/parameters/AchievementMaturity"
        - $ref: "#/components/parameters/ContentSortBy"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedAchievementSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /achievements:
    post:
      tags: [Achievements]
      summary: 创建成果草稿（发布方）
      operationId: createAchievement
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AchievementCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Achievement"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags: [Achievements]
      summary: 我的成果列表（发布方）
      operationId: listMyAchievements
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ContentStatus"
        - $ref: "#/components/parameters/AuditStatus"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedAchievement"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /achievements/{achievementId}:
    get:
      tags: [Achievements]
      summary: 获取成果详情（发布方/已登录）
      operationId: getAchievementById
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AchievementId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Achievement"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      tags: [Achievements]
      summary: 更新成果草稿（发布方）
      operationId: updateAchievement
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AchievementId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AchievementUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Achievement"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /achievements/{achievementId}/submit:
    post:
      tags: [Achievements]
      summary: 提交审核（发布方）
      operationId: submitAchievement
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AchievementId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Achievement"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /achievements/{achievementId}/off-shelf:
    post:
      tags: [Achievements]
      summary: 下架（发布方）
      operationId: offShelfAchievement
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AchievementId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  maxLength: 200
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Achievement"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/achievements:
    get:
      tags: [Admin]
      summary: 成果审核/平台内容列表
      operationId: adminListAchievementsForAudit
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Q"
        - $ref: "#/components/parameters/RegionCode"
        - $ref: "#/components/parameters/AuditStatus"
        - $ref: "#/components/parameters/ContentStatus"
        - $ref: "#/components/parameters/ContentSource"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedAchievement"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Admin]
      summary: 创建平台自有成果（P1）
      operationId: adminCreateAchievement
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAchievementCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Achievement"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/achievements/{achievementId}:
    get:
      tags: [Admin]
      summary: 获取成果详情（后台，P1）
      operationId: adminGetAchievementById
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AchievementId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Achievement"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Admin]
      summary: 更新平台自有成果（后台，P1）
      operationId: adminUpdateAchievement
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AchievementId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAchievementUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Achievement"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/achievements/{achievementId}/publish:
    post:
      tags: [Admin]
      summary: 发布成果（后台，P1）
      operationId: adminPublishAchievement
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AchievementId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Achievement"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /admin/achievements/{achievementId}/off-shelf:
    post:
      tags: [Admin]
      summary: 下架成果（后台，P1）
      operationId: adminOffShelfAchievement
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AchievementId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  maxLength: 200
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Achievement"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/achievements/{achievementId}/approve:
    post:
      tags: [Admin]
      summary: 成果审核通过
      operationId: adminApproveAchievement
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AchievementId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Achievement"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /admin/achievements/{achievementId}/reject:
    post:
      tags: [Admin]
      summary: 成果审核驳回
      operationId: adminRejectAchievement
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AchievementId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 1
                  maxLength: 500
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Achievement"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  parameters:
    Q:
      name: q
      in: query
      required: false
      schema:
        type: string
      description: 关键词（标题/摘要/权利人/发明人/机构名称等）
...
    RegionCode:
      name: regionCode
      in: query
      required: false
      schema:
        type: string
        pattern: "^[0-9]{6}$"
        example: "110000"
      description: 行政区划 adcode（6 位字符串）
...
    IndustryTags:
      name: industryTags
      in: query
      required: false
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
      description: 产业标签过滤（可多选）
...
    ContentSortBy:
      name: sortBy
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/ContentSortBy"
...
    CooperationModes:
      name: cooperationModes
      in: query
      required: false
      schema:
        type: array
        items:
          $ref: "#/components/schemas/CooperationMode"
      style: form
      explode: true
      description: 合作方式过滤（可多选）
    AchievementMaturity:
      name: maturity
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/AchievementMaturity"
    ContentStatus:
      name: status
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/ContentStatus"
...
    ContentSource:
      name: source
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/ContentSource"
      description: 内容来源过滤
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    AuditStatus:
      name: auditStatus
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/AuditStatus"
...
    AchievementId:
      name: achievementId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
  schemas:
    Uuid:
      type: string
      format: uuid
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    VerificationType:
      type: string
      enum: [PERSON, COMPANY, ACADEMY, GOVERNMENT, ASSOCIATION, TECH_MANAGER]
    SupplyType:
      type: string
      enum: [UNIVERSITY, UNIVERSITY_985, UNIVERSITY_211, RESEARCH_INSTITUTE, OTHER]
    VerificationStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
...
    AuditStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
    ContentSource:
      type: string
      enum: [USER, PLATFORM, ADMIN]
...
    ListingStats:
      type: object
      required: [viewCount, favoriteCount, consultCount]
      properties:
        viewCount:
          type: integer
          format: int64
          minimum: 0
        favoriteCount:
          type: integer
          format: int64
          minimum: 0
        consultCount:
          type: integer
          format: int64
          minimum: 0
        commentCount:
          type: integer
          format: int64
          minimum: 0
...
    PageMeta:
      type: object
      required: [page, pageSize, total]
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
...
    ContentStatus:
      type: string
      enum: [DRAFT, ACTIVE, OFF_SHELF]
    ContentSortBy:
      type: string
      enum: [RECOMMENDED, NEWEST]
...
    CooperationMode:
      type: string
      enum: [TRANSFER, TECH_CONSULTING, COMMISSIONED_DEV, PLATFORM_CO_BUILD]
    AchievementMaturity:
      type: string
      enum: [CONCEPT, PROTOTYPE, PILOT, MASS_PRODUCTION, COMMERCIALIZED, OTHER]
...
    ContentMedia:
      type: object
      required: [fileId, type, sort]
      properties:
        fileId:
          $ref: "#/components/schemas/Uuid"
        type:
          type: string
          enum: [IMAGE, VIDEO, FILE]
        sort:
          type: integer
          minimum: 0
        url:
          type: string
          format: uri
          readOnly: true
        mimeType:
          type: string
          readOnly: true
        sizeBytes:
          type: integer
          format: int64
          readOnly: true
        fileName:
          type: string
          readOnly: true
...
    AchievementSummary:
      type: object
      required: [id, title, publisher, auditStatus, status, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        source:
          $ref: "#/components/schemas/ContentSource"
        title:
          type: string
          maxLength: 200
        summary:
          type: string
          maxLength: 2000
        maturity:
          $ref: "#/components/schemas/AchievementMaturity"
        cooperationModes:
          type: array
          items:
            $ref: "#/components/schemas/CooperationMode"
        regionCode:
          type: string
        industryTags:
          type: array
          items:
            type: string
        keywords:
          type: array
          maxItems: 30
          items:
            type: string
            maxLength: 50
        publisher:
          $ref: "#/components/schemas/OrganizationSummary"
        stats:
          $ref: "#/components/schemas/ListingStats"
        auditStatus:
          $ref: "#/components/schemas/AuditStatus"
        status:
          $ref: "#/components/schemas/ContentStatus"
        coverUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
    Achievement:
      allOf:
        - $ref: "#/components/schemas/AchievementSummary"
        - type: object
          properties:
            publisherUserId:
              $ref: "#/components/schemas/Uuid"
            description:
              type: string
              maxLength: 5000
            coverFileId:
              $ref: "#/components/schemas/Uuid"
              nullable: true
            media:
              type: array
              items:
                $ref: "#/components/schemas/ContentMedia"
            aiParse:
              $ref: "#/components/schemas/AiParseResult"
              nullable: true
            updatedAt:
              type: string
              format: date-time
    AchievementPublic:
      allOf:
        - $ref: "#/components/schemas/AchievementSummary"
        - type: object
          properties:
            description:
              type: string
              maxLength: 5000
            media:
              type: array
              items:
                $ref: "#/components/schemas/ContentMedia"
            aiParse:
              $ref: "#/components/schemas/AiParseResult"
              nullable: true
    PagedAchievementSummary:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AchievementSummary"
        page:
          $ref: "#/components/schemas/PageMeta"
    PagedAchievement:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Achievement"
        page:
          $ref: "#/components/schemas/PageMeta"
    AchievementCreateRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 5000
        summary:
          type: string
          maxLength: 2000
        keywords:
          type: array
          maxItems: 30
          items:
            type: string
            maxLength: 50
        maturity:
          $ref: "#/components/schemas/AchievementMaturity"
        cooperationModes:
          type: array
          items:
            $ref: "#/components/schemas/CooperationMode"
        regionCode:
          type: string
        industryTags:
          type: array
          items:
            type: string
        coverFileId:
          $ref: "#/components/schemas/Uuid"
          nullable: true
        media:
          type: array
          items:
            $ref: "#/components/schemas/ContentMedia"
    AchievementUpdateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 5000
        summary:
          type: string
          maxLength: 2000
        keywords:
          type: array
          maxItems: 30
          items:
            type: string
            maxLength: 50
        maturity:
          $ref: "#/components/schemas/AchievementMaturity"
        cooperationModes:
          type: array
          items:
            $ref: "#/components/schemas/CooperationMode"
        regionCode:
          type: string
        industryTags:
          type: array
          items:
            type: string
        coverFileId:
          $ref: "#/components/schemas/Uuid"
          nullable: true
        media:
          type: array
          items:
            $ref: "#/components/schemas/ContentMedia"
...
    OrganizationStats:
      type: object
      required: [listingCount, patentCount]
      properties:
        listingCount:
          type: integer
          minimum: 0
        patentCount:
          type: integer
          minimum: 0
    OrganizationSummary:
      type: object
      required: [userId, displayName, verificationType, verificationStatus]
      properties:
        userId:
          $ref: "#/components/schemas/Uuid"
        displayName:
          type: string
        verificationType:
          $ref: "#/components/schemas/VerificationType"
        verificationStatus:
          $ref: "#/components/schemas/VerificationStatus"
        orgCategory:
          $ref: "#/components/schemas/SupplyType"
        regionCode:
          type: string
        logoUrl:
          type: string
          format: uri
        intro:
          type: string
        stats:
          $ref: "#/components/schemas/OrganizationStats"
        verifiedAt:
          type: string
          format: date-time
...
    AiContentType:
      type: string
      enum: [LISTING, DEMAND, ACHIEVEMENT, ARTWORK]
    AiParseStatus:
      type: string
      enum: [ACTIVE, REVIEW_REQUIRED, REPLACED]
...
    AiParseResult:
      type: object
      required: [id, contentType, contentId, status, confidence, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        contentType:
          $ref: "#/components/schemas/AiContentType"
        contentId:
          $ref: "#/components/schemas/Uuid"
        summaryPlain:
          type: string
        featuresPlain:
          type: string
        keywords:
          type: array
          items:
            type: string
        confidence:
          type: number
          format: double
        modelVersion:
          type: string
        status:
          $ref: "#/components/schemas/AiParseStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AdminAchievementCreateRequest:
      allOf:
        - $ref: "#/components/schemas/AchievementCreateRequest"
        - type: object
          properties:
            source:
              $ref: "#/components/schemas/ContentSource"
            publisherUserId:
              $ref: "#/components/schemas/Uuid"
            auditStatus:
              $ref: "#/components/schemas/AuditStatus"
            status:
              $ref: "#/components/schemas/ContentStatus"
    AdminAchievementUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/AchievementUpdateRequest"
        - type: object
          properties:
            source:
              $ref: "#/components/schemas/ContentSource"
            publisherUserId:
              $ref: "#/components/schemas/Uuid"
            auditStatus:
              $ref: "#/components/schemas/AuditStatus"
            status:
              $ref: "#/components/schemas/ContentStatus"
...